move CollectPlayerInput function from PlayerInput to  Player
