--- Tic-Tac-Toe/IPrintBoard.cs
@@ -1,7 +1,0 @@
-namespace Tic_Tac_Toe
-{
-    public interface IPrintBoard
-    {
-        void Print();
-    }
-}
