+++ Tic-Tac-Toe/Player.cs
@@ -1,3 +1,4 @@
+using System;
 namespace Tic_Tac_Toe
 {
     public class Player
@@ -9,6 +10,19 @@ namespace Tic_Tac_Toe
         {
             CellValue = cellValue;
             Name = name;
+        }
+        
+        private string Ask(string question)
+        {
+            Console.Write(question);
+            return Console.ReadLine();
+        }
+        
+        public string CollectPlayerInput()
+        {
+            var instruction = $"{Name} enter a coord x,y to place your {CellValue} or enter 'q' to give up: ";
+            var input = Ask(instruction);
+            return input;
         }
     }
